new files added

+++ zanTech/app/Http/Controllers/authController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Product;
@@ -9,6 +10,7 @@ use Illuminate\Http\Request;
 use ShoppingCart;
 use RealRashid\SweetAlert\Facades\Alert;
 use Illuminate\Support\Facades\Hash;
+
 class authController extends Controller
 {
     public function ViewLogin()
@@ -26,44 +28,46 @@ class authController extends Controller
         $userType = Auth::user()->userType;
 
         if ($userType == 'Admin') {
-            return view('admin.adminDashbaoard');
+            return view('admin.adminDashboard');
         }
 
         if ($userType == 0) {
-            // all product
-            $products = Product::all();
+            // all product withour projects
+            $products = Product::where('catagory', '!=', 'Project')->orderBy('id','desc')->get();
             //last 10 uploaded product
-            $latestProducts = Product::latest()->take(10)->get();
+            $latestProducts = Product::where('catagory', '!=', 'Project')->latest()->take(10)->get();
             // most ordered product
             $trendingProductCounts = DB::table('order_items')
-            ->select('productName', DB::raw('COUNT(*) as order_count'))
-            ->groupBy('productName')
-            ->orderBy('order_count', 'asc')
-            ->limit(15)
-            ->pluck('productName');
-        $trendingProducts = Product::whereIn('productName', $trendingProductCounts)
-        ->get();
+                ->select('productName', DB::raw('COUNT(*) as order_count'))
+                ->groupBy('productName')
+                ->orderBy('order_count', 'desc')
+                ->limit(15)
+                ->pluck('productName');
+            $trendingProductCounts = $trendingProductCounts->toArray();
 
-            return view('user.userDashboard', compact('latestProducts','products', 'trendingProducts'));
+            $trendingProducts = Product::whereIn('productName', $trendingProductCounts)
+                ->get()
+                ->sortBy(function ($product) use ($trendingProductCounts) {
+                    return array_search($product->productName, $trendingProductCounts);
+                });
 
+            return view('user.userDashboard', compact('latestProducts', 'products', 'trendingProducts'));
         } else {
             return View('auth.login');
         }
     }
 
     public function logoutFunction()
-    {if (Auth::check()) {
-        if (Auth::user()->userType == "0") {
-            Auth::logout();
-            ShoppingCart::destroy();
-            return Redirect('/');
-
+    {
+        if (Auth::check()) {
+            if (Auth::user()->userType == "0") {
+                Auth::logout();
+                ShoppingCart::destroy();
+                return Redirect('/');
+            } else
+                Auth::logout();
+            return Redirect()->route('login');
         }
-        else
-
-        Auth::logout();
-        return Redirect()->route('login');
-    }
     }
     public function userProfile()
     {
@@ -74,7 +78,6 @@ class authController extends Controller
 
             return view('auth.login');
         }
-
     }
     public function updateInfo(Request $request)
     {
@@ -85,7 +88,7 @@ class authController extends Controller
             if (!$updateInfo) {
                 return redirect()->back()->with('error', 'ID Not found');
             }
-            if($updateInfo){
+            if ($updateInfo) {
                 $updateInfo->name = $request->input('updateName');
                 $updateInfo->email = $request->input('updateEmail');
                 $updateInfo->phone = $request->input('updatePhone');
@@ -93,7 +96,6 @@ class authController extends Controller
                 Alert::success('Successfull', 'Profile updated');
                 return redirect()->back();
             }
-
         } else {
 
             return view('auth.login');
@@ -108,25 +110,20 @@ class authController extends Controller
             if (!$updateInfo) {
                 return redirect()->back()->with('error', 'ID Not found');
             }
-            if (Hash::check($request->currentPassword, $updateInfo ->password)) {
-                if($request->NewPassword== $request->ConfirmPassword){
-                    // $updateInfo->password = password_hash($request->newPassword, PASSWORD_DEFAULT);
+            if (Hash::check($request->currentPassword, $updateInfo->password)) {
+                if ($request->NewPassword == $request->ConfirmPassword) {
                     $updateInfo->password = Hash::make($request->input('NewPassword'));
                     $updateInfo->save();
                     Alert::success('success', 'Password changed successfully');
                     return redirect()->back();
-                }
-                else{
+                } else {
                     Alert::warning('error', 'password does not match!!');
                     return redirect()->back();
                 }
-
             } else {
-                // If the current password doesn't match, return back with an error message
                 Alert::warning('error', 'Current password is incorrect');
                 return redirect()->back();
             }
-
         } else {
 
             return view('auth.login');
